Patch 0.1.2: Updated all repository section into an accordion
